Refactor handlers into tab-specific modules
Split handlers.py (~800 lines) into modular structure:
- handlers.py now contains only shared helpers (~80 lines)
- tabs/battery/handlers.py: battery-specific handlers
- tabs/compare/handlers.py: compare-specific handlers

This improves maintainability and prepares for future tab additions.
Updated ARCHITECTURE.md to document the new module structure.
