+++ prompt_prix/ui.py
@@ -3,6 +3,8 @@ Gradio UI definition for prompt-prix.
 
 Battery-first design: Model × Test grid is the primary interface.
 Interactive comparison is secondary (Compare tab).
+
+Tab-specific handlers are in prompt_prix.tabs.{battery,compare}.handlers
 """
 
 import gradio as gr
@@ -14,22 +16,7 @@ from prompt_prix.config import (
     DEFAULT_TIMEOUT_SECONDS,
     DEFAULT_MAX_TOKENS
 )
-from prompt_prix.handlers import (
-    fetch_available_models,
-    initialize_session,
-    send_single_prompt,
-    export_markdown,
-    export_json,
-    battery_validate_file,
-    battery_run_handler,
-    battery_get_cell_detail,
-    battery_get_test_ids,
-    battery_quick_prompt_handler,
-    battery_export_json,
-    battery_export_csv,
-    battery_refresh_grid,
-    handle_stop
-)
+from prompt_prix.handlers import fetch_available_models, handle_stop
 from prompt_prix.parsers import get_default_system_prompt
 from prompt_prix.ui_helpers import (
     CUSTOM_CSS,
@@ -39,10 +26,10 @@ from prompt_prix.ui_helpers import (
     SAVE_TEMPERATURE_JS,
 )
 
+# Import tab-specific handlers
+from prompt_prix.tabs.battery import handlers as battery_handlers
+from prompt_prix.tabs.compare import handlers as compare_handlers
 
-# ─────────────────────────────────────────────────────────────────────
-# APPLICATION
-# ─────────────────────────────────────────────────────────────────────
 
 def create_app() -> gr.Blocks:
     """Create the Gradio application."""
@@ -53,12 +40,11 @@ def create_app() -> gr.Blocks:
     blocks_kwargs = {"title": "prompt-prix"}
 
     if "theme" in blocks_params:
-        # Gradio 5.x style
         blocks_kwargs["theme"] = gr.themes.Soft()
         blocks_kwargs["css"] = CUSTOM_CSS
 
     with gr.Blocks(**blocks_kwargs) as app:
-        
+
         gr.Markdown("# prompt-prix")
         gr.Markdown("Find your optimal open-weights model.")
 
@@ -84,8 +70,7 @@ def create_app() -> gr.Blocks:
                         interactive=False,
                         lines=2
                     )
-        
-        # Shared model list (populated by fetch, used by both tabs)
+
         available_models = gr.State([])
 
         # ─────────────────────────────────────────────────────────────
@@ -93,20 +78,19 @@ def create_app() -> gr.Blocks:
         # ─────────────────────────────────────────────────────────────
 
         with gr.Tabs() as main_tabs:
-            
+
             # ─────────────────────────────────────────────────────────
             # TAB 1: BATTERY (Primary)
             # ─────────────────────────────────────────────────────────
-            
+
             with gr.Tab("🔋 Battery", id="battery-tab"):
-                
+
                 gr.Markdown("""
-                Run benchmark test suites across multiple models. 
+                Run benchmark test suites across multiple models.
                 Upload a test file, select models, and see results in the grid below.
                 """)
-                
+
                 with gr.Row():
-                    # Left column: Test suite + Models
                     with gr.Column(scale=1):
                         battery_file = gr.File(
                             label="Test Suite (JSON/JSONL)",
@@ -119,7 +103,7 @@ def create_app() -> gr.Blocks:
                             interactive=False,
                             lines=1
                         )
-                        
+
                         with gr.Row():
                             battery_models = gr.CheckboxGroup(
                                 label="Models to Test",
@@ -134,8 +118,7 @@ def create_app() -> gr.Blocks:
                                 size="sm",
                                 scale=1
                             )
-                    
-                    # Right column: Config
+
                     with gr.Column(scale=1):
                         battery_temp = gr.Slider(
                             label="Temperature",
@@ -164,7 +147,6 @@ def create_app() -> gr.Blocks:
                             lines=3
                         )
 
-                        # Quick prompt for ad-hoc testing
                         gr.Markdown("---")
                         gr.Markdown("**Quick Prompt**")
                         quick_prompt = gr.Textbox(
@@ -194,16 +176,12 @@ def create_app() -> gr.Blocks:
                         variant="stop",
                         scale=1
                     )
-                
+
                 battery_status = gr.Textbox(
                     label="Status",
                     value="Ready",
                     interactive=False
                 )
-                
-                # ─────────────────────────────────────────────────────
-                # HERO: Results Grid
-                # ─────────────────────────────────────────────────────
 
                 gr.Markdown("### Results")
                 battery_display_mode = gr.Radio(
@@ -220,8 +198,7 @@ def create_app() -> gr.Blocks:
                     wrap=True,
                     elem_id="battery-grid"
                 )
-                
-                # Detail view for clicked cell
+
                 with gr.Accordion("📋 Response Detail", open=False):
                     with gr.Row():
                         detail_model = gr.Dropdown(
@@ -235,12 +212,11 @@ def create_app() -> gr.Blocks:
                             interactive=True
                         )
                         detail_refresh_btn = gr.Button("Show", size="sm")
-                    
+
                     detail_content = gr.Markdown(
                         value="*Select a model and test to view the response*"
                     )
-                
-                # Export
+
                 with gr.Row():
                     battery_export_json_btn = gr.Button("Export JSON")
                     battery_export_csv_btn = gr.Button("Export CSV")
@@ -255,7 +231,7 @@ def create_app() -> gr.Blocks:
             # ─────────────────────────────────────────────────────────
             # TAB 2: COMPARE (Secondary)
             # ─────────────────────────────────────────────────────────
-            
+
             with gr.Tab("💬 Compare", id="compare-tab"):
 
                 gr.Markdown("""
@@ -264,7 +240,6 @@ def create_app() -> gr.Blocks:
                 """)
 
                 with gr.Row():
-                    # Left column: Models
                     with gr.Column(scale=1):
                         with gr.Row():
                             compare_models = gr.CheckboxGroup(
@@ -281,7 +256,6 @@ def create_app() -> gr.Blocks:
                                 scale=1
                             )
 
-                    # Right column: Config + Quick Prompt
                     with gr.Column(scale=1):
                         compare_temp = gr.Slider(
                             label="Temperature",
@@ -320,7 +294,6 @@ def create_app() -> gr.Blocks:
                                 elem_id="tools"
                             )
 
-                        # Quick prompt for comparison
                         gr.Markdown("---")
                         gr.Markdown("**Prompt**")
                         compare_prompt = gr.Textbox(
@@ -338,23 +311,20 @@ def create_app() -> gr.Blocks:
                     value="Select models and send a prompt",
                     interactive=False
                 )
-                
-                # Hidden state for tab colors
+
                 tab_states = gr.JSON(value=[], visible=False, elem_id="tab-states")
-                
-                # Model output tabs (up to 10)
+
                 model_outputs = []
                 with gr.Tabs(elem_id="model-tabs"):
                     for i in range(10):
                         with gr.Tab(f"Model {i + 1}"):
                             output = gr.Markdown(value="", label="Conversation")
                             model_outputs.append(output)
-                
-                # Export
+
                 with gr.Row():
                     compare_export_md_btn = gr.Button("Export Markdown")
                     compare_export_json_btn = gr.Button("Export JSON")
-                
+
                 compare_export_preview = gr.Textbox(
                     label="Export Preview",
                     lines=10,
@@ -372,47 +342,28 @@ def create_app() -> gr.Blocks:
             choices = models_update.get("choices", []) if isinstance(models_update, dict) else []
             return (
                 status,
-                choices,  # State
-                gr.update(choices=choices),  # Battery models
-                gr.update(choices=choices),  # Compare models
-                gr.update(choices=choices),  # Detail model dropdown
+                choices,
+                gr.update(choices=choices),
+                gr.update(choices=choices),
+                gr.update(choices=choices),
             )
 
         fetch_models_btn.click(
             fn=on_fetch_models,
             inputs=[servers_input],
-            outputs=[
-                server_status,
-                available_models,
-                battery_models,
-                compare_models,
-                detail_model
-            ]
+            outputs=[server_status, available_models, battery_models, compare_models, detail_model]
         )
 
-        # Tab-local fetch buttons (same handler, updates both tabs)
         battery_fetch_btn.click(
             fn=on_fetch_models,
             inputs=[servers_input],
-            outputs=[
-                server_status,
-                available_models,
-                battery_models,
-                compare_models,
-                detail_model
-            ]
+            outputs=[server_status, available_models, battery_models, compare_models, detail_model]
         )
 
         compare_fetch_btn.click(
             fn=on_fetch_models,
             inputs=[servers_input],
-            outputs=[
-                server_status,
-                available_models,
-                battery_models,
-                compare_models,
-                detail_model
-            ]
+            outputs=[server_status, available_models, battery_models, compare_models, detail_model]
         )
 
         # ─────────────────────────────────────────────────────────────
@@ -420,15 +371,12 @@ def create_app() -> gr.Blocks:
         # ─────────────────────────────────────────────────────────────
 
         def on_battery_file_change(file_obj):
-            """Validate file and enable/disable run button."""
             if file_obj is None:
                 return "Upload a benchmark file", gr.update(interactive=False), gr.update(choices=[])
 
-            validation_result = battery_validate_file(file_obj)
+            validation_result = battery_handlers.validate_file(file_obj)
             is_valid = validation_result.startswith("✅")
-
-            # Extract test IDs for detail dropdown (if valid)
-            test_choices = battery_get_test_ids(file_obj) if is_valid else []
+            test_choices = battery_handlers.get_test_ids(file_obj) if is_valid else []
 
             return (
                 validation_result,
@@ -443,72 +391,47 @@ def create_app() -> gr.Blocks:
         )
 
         battery_run_btn.click(
-            fn=battery_run_handler,
+            fn=battery_handlers.run_handler,
             inputs=[
-                battery_file,
-                battery_models,
-                servers_input,
-                battery_temp,
-                battery_timeout,
-                battery_max_tokens,
-                battery_system_prompt
+                battery_file, battery_models, servers_input,
+                battery_temp, battery_timeout, battery_max_tokens, battery_system_prompt
             ],
             outputs=[battery_status, battery_grid]
         )
 
-        # Quick prompt
         quick_prompt_btn.click(
-            fn=battery_quick_prompt_handler,
+            fn=battery_handlers.quick_prompt_handler,
             inputs=[
-                quick_prompt,
-                battery_models,
-                servers_input,
-                battery_temp,
-                battery_timeout,
-                battery_max_tokens,
-                battery_system_prompt
+                quick_prompt, battery_models, servers_input,
+                battery_temp, battery_timeout, battery_max_tokens, battery_system_prompt
             ],
             outputs=[quick_prompt_output]
         )
 
-        # Stop button - signals cancellation to running handlers
-        battery_stop_btn.click(
-            fn=handle_stop,
-            inputs=[],
-            outputs=[battery_status]
-        )
+        battery_stop_btn.click(fn=handle_stop, inputs=[], outputs=[battery_status])
 
-        # Display mode toggle - refresh grid with selected mode
         battery_display_mode.change(
-            fn=battery_refresh_grid,
+            fn=battery_handlers.refresh_grid,
             inputs=[battery_display_mode],
             outputs=[battery_grid]
         )
 
-        # Detail view
         detail_refresh_btn.click(
-            fn=battery_get_cell_detail,
+            fn=battery_handlers.get_cell_detail,
             inputs=[detail_model, detail_test],
             outputs=[detail_content]
         )
 
-        # Battery export
         battery_export_json_btn.click(
-            fn=battery_export_json,
+            fn=battery_handlers.export_json,
             inputs=[],
             outputs=[battery_status, battery_export_preview]
-        ).then(
-            fn=lambda: gr.update(visible=True),
-            outputs=[battery_export_preview]
         )
 
         battery_export_csv_btn.click(
-            fn=battery_export_csv,
+            fn=battery_handlers.export_csv,
             inputs=[],
             outputs=[battery_status, battery_export_preview]
-        ).then(
-            fn=lambda: gr.update(visible=True),
-            outputs=[battery_export_preview]
         )
 
         # ─────────────────────────────────────────────────────────────
@@ -519,21 +442,17 @@ def create_app() -> gr.Blocks:
             prompt, tools, servers_text, models_selected,
             system_prompt, temperature, timeout, max_tokens
         ):
-            """Auto-initialize session if needed, then send prompt."""
-            # Auto-initialize if session doesn't exist or models changed
             if (state.session is None or
                 set(state.session.state.models) != set(models_selected)):
-                init_status, *init_outputs = await initialize_session(
+                init_status, *init_outputs = await compare_handlers.initialize_session(
                     servers_text, models_selected, system_prompt,
                     temperature, timeout, max_tokens
                 )
                 if "❌" in init_status or "⚠️" in init_status:
-                    # Initialization failed
                     yield (init_status, []) + tuple(init_outputs)
                     return
 
-            # Now send the prompt
-            async for result in send_single_prompt(prompt, tools):
+            async for result in compare_handlers.send_single_prompt(prompt, tools):
                 yield result
 
         compare_send_btn.click(
@@ -554,36 +473,24 @@ def create_app() -> gr.Blocks:
             outputs=[compare_status, tab_states] + model_outputs
         )
 
-        tab_states.change(
-            fn=None,
-            inputs=[tab_states],
-            outputs=[tab_states],
-            js=TAB_STATUS_JS
-        )
+        tab_states.change(fn=None, inputs=[tab_states], outputs=[tab_states], js=TAB_STATUS_JS)
 
         compare_export_md_btn.click(
-            fn=export_markdown,
+            fn=compare_handlers.export_markdown,
             inputs=[],
             outputs=[compare_status, compare_export_preview]
-        ).then(
-            fn=lambda: gr.update(visible=True),
-            outputs=[compare_export_preview]
-        )
+        ).then(fn=lambda: gr.update(visible=True), outputs=[compare_export_preview])
 
         compare_export_json_btn.click(
-            fn=export_json,
+            fn=compare_handlers.export_json,
             inputs=[],
             outputs=[compare_status, compare_export_preview]
-        ).then(
-            fn=lambda: gr.update(visible=True),
-            outputs=[compare_export_preview]
-        )
+        ).then(fn=lambda: gr.update(visible=True), outputs=[compare_export_preview])
 
         # ─────────────────────────────────────────────────────────────
-        # PERSISTENCE: Load from localStorage, save on change
+        # PERSISTENCE
         # ─────────────────────────────────────────────────────────────
 
-        # Load saved settings (returns undefined to preserve .env defaults if empty)
         app.load(
             fn=None,
             inputs=[],
@@ -591,26 +498,8 @@ def create_app() -> gr.Blocks:
             js=PERSISTENCE_LOAD_JS
         )
 
-        # Save servers to localStorage on change (debounced by Gradio for textboxes)
-        servers_input.change(
-            fn=None,
-            inputs=[servers_input],
-            outputs=[servers_input],
-            js=SAVE_SERVERS_JS
-        )
-
-        # Save temperature to localStorage on change (syncs both sliders)
-        battery_temp.change(
-            fn=None,
-            inputs=[battery_temp],
-            outputs=[battery_temp],
-            js=SAVE_TEMPERATURE_JS
-        )
-        compare_temp.change(
-            fn=None,
-            inputs=[compare_temp],
-            outputs=[compare_temp],
-            js=SAVE_TEMPERATURE_JS
-        )
+        servers_input.change(fn=None, inputs=[servers_input], outputs=[servers_input], js=SAVE_SERVERS_JS)
+        battery_temp.change(fn=None, inputs=[battery_temp], outputs=[battery_temp], js=SAVE_TEMPERATURE_JS)
+        compare_temp.change(fn=None, inputs=[compare_temp], outputs=[compare_temp], js=SAVE_TEMPERATURE_JS)
 
     return app
